added AI bots for singleplayer mode
The easy bot tries to play 1 card first, then 2 cards and then 3 cards.
The hard bot tries to play 3 cards first, then 2 cards and then 1 card.
Also fixed a big bug on  player card checking that wasn't apparent until now,
which mismatched the info on what cards players had in the frontend and in the
backend.